+++ python/egglog/__init__.py
@@ -4,6 +4,7 @@ Package for creating e-graphs in Python.
 
 from . import config, ipython_magic  # noqa: F401
 from .bindings import EggSmolError, StageInfo, TimeOnly, WithPlan  # noqa: F401
+from .bindings import Program as SharedProgram  # noqa: F401
 from .builtins import *
 from .conversion import *
 from .deconstruct import *